+++ yash-builtin/CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `InvalidVariableName` variant, which indicates that the variable name is
   invalid.
 
+### Changed
+
+- External dependency versions:
+    - yash-syntax 0.14.0 → 0.14.1
+
 ### Fixed
 
 - The `getopts` built-in now fails when the second operand is not a valid
@@ -24,6 +29,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The `export`, `readonly`, and `typeset` built-ins no longer print variables
   with a name containing an `=` character.
   The `typeset::PrintVariables::execute` function now ignores such variables.
+- The `set` built-in without arguments no longer prints variables that have an
+  invalid name. The `set::main` function now excludes such variables from the
+  output.
 
 ## [0.7.0] - 2025-04-26